I'm still seeing a CallbackStack leak at $work with the most recent 3.5.3 release 😓 I chatted it over with Arman, and we think it's the situation he outlined here: https://github.com/typelevel/cats-effect/pull/3943#discussion_r1448077903

This PR updates `apply` to clean up the stack as it goes, ~and the `clear` method to null everything out much more aggressively~ Actually, I've removed the `clear` method on the stack, and added a `try` / `finally` to the apply so that `apply` will finish tearing down if a callback throws an exception. This is necessary because by tearing down as `apply` goes, `clear` would have no way to access the node that `apply` got to, as the links between `head` and that node would have been broken.