@durban 2024-01-27
Btw, if a callback throws an exception, why aren't we calling the rest of the callbacks? (I realize this is also the current behavior.) I see @armanbilge also already asked this [here](https://github.com/typelevel/cats-effect/pull/3943#discussion_r1451633273).