@djspiewak 2024-01-27
> Btw, if a callback throws an exception, why aren't we calling the rest of the callbacks?

Oh that very much seems like a bug. It's harmless right now since `cont`'s `resume` will never throw, but I would prefer to be defensive about it unless it costs us performance to do so.