@djspiewak 2024-01-27
I'll run some benchmarks to sanity check.