@djspiewak 2024-01-28
Essentially a wash on performance, at least so far as our benchmarks can measure.

## Before

```
[info] Benchmark                                             (count)  (cpuTokens)   (size)   Mode  Cnt     Score    Error    Units
[info] DeferredBenchmark.cancel                                   10          N/A      N/A  thrpt   10   110.427 ±  0.496    ops/s
[info] DeferredBenchmark.cancel                                  100          N/A      N/A  thrpt   10    14.109 ±  0.055    ops/s
[info] DeferredBenchmark.cancel                                 1000          N/A      N/A  thrpt   10     1.318 ±  0.005    ops/s
[info] DeferredBenchmark.complete                                 10          N/A      N/A  thrpt   10   108.645 ±  0.414    ops/s
[info] DeferredBenchmark.complete                                100          N/A      N/A  thrpt   10    25.782 ±  0.136    ops/s
[info] DeferredBenchmark.complete                               1000          N/A      N/A  thrpt   10     5.018 ±  0.245    ops/s
[info] DeferredBenchmark.getAfter                                 10          N/A      N/A  thrpt   10  6915.081 ± 24.629    ops/s
[info] DeferredBenchmark.getAfter                                100          N/A      N/A  thrpt   10  1929.931 ±  7.287    ops/s
[info] DeferredBenchmark.getAfter                               1000          N/A      N/A  thrpt   10   230.685 ±  1.069    ops/s
[info] DeferredBenchmark.getBefore                                10          N/A      N/A  thrpt   10  3929.464 ±  5.893    ops/s
[info] DeferredBenchmark.getBefore                               100          N/A      N/A  thrpt   10   745.665 ±  1.358    ops/s
[info] DeferredBenchmark.getBefore                              1000          N/A      N/A  thrpt   10    86.061 ±  0.183    ops/s
[info] ParallelBenchmark.parTraverse                             N/A        10000     1000  thrpt   10   875.136 ± 24.879    ops/s
[info] ParallelBenchmark.traverse                                N/A        10000     1000  thrpt   10    70.208 ±  0.165    ops/s
[info] WorkStealingBenchmark.alloc                               N/A          N/A  1000000  thrpt   10    13.987 ±  0.173  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark      N/A          N/A  1000000  thrpt   10    48.534 ±  2.015  ops/min
[info] WorkStealingBenchmark.runnableScheduling                  N/A          N/A  1000000  thrpt   10  2972.862 ± 20.806  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal       N/A          N/A  1000000  thrpt   10  2229.858 ± 10.108  ops/min
[info] WorkStealingBenchmark.scheduling                          N/A          N/A  1000000  thrpt   10    49.855 ±  3.070  ops/min
```

## After

```
[info] Benchmark                                             (count)  (cpuTokens)   (size)   Mode  Cnt     Score    Error    Units
[info] DeferredBenchmark.cancel                                   10          N/A      N/A  thrpt   10   111.104 ±  0.618    ops/s
[info] DeferredBenchmark.cancel                                  100          N/A      N/A  thrpt   10    13.589 ±  0.070    ops/s
[info] DeferredBenchmark.cancel                                 1000          N/A      N/A  thrpt   10     1.413 ±  0.007    ops/s
[info] DeferredBenchmark.complete                                 10          N/A      N/A  thrpt   10   107.398 ±  0.614    ops/s
[info] DeferredBenchmark.complete                                100          N/A      N/A  thrpt   10    25.566 ±  0.153    ops/s
[info] DeferredBenchmark.complete                               1000          N/A      N/A  thrpt   10     5.072 ±  0.249    ops/s
[info] DeferredBenchmark.getAfter                                 10          N/A      N/A  thrpt   10  6792.078 ± 57.712    ops/s
[info] DeferredBenchmark.getAfter                                100          N/A      N/A  thrpt   10  1920.488 ±  8.127    ops/s
[info] DeferredBenchmark.getAfter                               1000          N/A      N/A  thrpt   10   231.931 ±  0.958    ops/s
[info] DeferredBenchmark.getBefore                                10          N/A      N/A  thrpt   10  3916.262 ±  7.243    ops/s
[info] DeferredBenchmark.getBefore                               100          N/A      N/A  thrpt   10   745.866 ±  1.469    ops/s
[info] DeferredBenchmark.getBefore                              1000          N/A      N/A  thrpt   10    85.573 ±  0.235    ops/s
[info] ParallelBenchmark.parTraverse                             N/A        10000     1000  thrpt   10   867.002 ± 25.050    ops/s
[info] ParallelBenchmark.traverse                                N/A        10000     1000  thrpt   10    70.059 ±  0.122    ops/s
[info] WorkStealingBenchmark.alloc                               N/A          N/A  1000000  thrpt   10    14.005 ±  0.094  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark      N/A          N/A  1000000  thrpt   10    49.554 ±  2.297  ops/min
[info] WorkStealingBenchmark.runnableScheduling                  N/A          N/A  1000000  thrpt   10  2730.891 ± 86.788  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal       N/A          N/A  1000000  thrpt   10  2228.855 ±  9.531  ops/min
[info] WorkStealingBenchmark.scheduling                          N/A          N/A  1000000  thrpt   10    50.375 ±  5.190  ops/min
```