@aartigao 2024-04-10
@samspills Did you test 3.5.4 to see if the leak was fixed?

We're having a leak too, and Eclipse MAT points to `CallbackStack.Node`. I'm trying to figure out if it's something we're doing wrong and the leak of `Node`s is just a bad usage of the CE API.

![image](https://github.com/typelevel/cats-effect/assets/6449814/ba9e6306-26e9-4ace-9821-1c48b3ce69bb)
![image](https://github.com/typelevel/cats-effect/assets/6449814/459d719e-ec90-4109-b26f-04b9c24a80fa)