@samspills 2024-04-10
@aartigao I tested a snapshot with this commit and the leak was _not_ fixed. I've been poking at it when I can but I haven't been able to dedicate myself to it (kubernetes shields us from most of the OOM pain currently)

I've found that my leak seems to be caused by IOFiber joins, and in that use case the CallbackStack is not `pack`ed the way it is in the Deferred case. To me this says that it is indeed a bad usage of the CE API with IOFiber, but I haven't had time to nail down where (no usage of IOFiber in my app directly). From your class histogram, can I guess that you're using fs2-kafka? My leaking app is also an fs2-kafka app. I have other fs2-kafka apps that are behaving well but if that is a commonality in our cases then that is interesting