@aartigao 2024-04-10
>From your class histogram, can I guess that you're using fs2-kafka?

Yup! I've been digging and our leak matches in time when our traffic descends to 0 and there are no more Kafka records to consume. It can be an `fs2-kafka` thing, but we're suspicious about a part of our code (some really complex code) that may be leaking fibers in the wild. In our case it's possible to misuse the CE Fiber API because that code makes use of `racePair` and might be not cancelling the loser fiber.

Let's see, we'll run another test this afternoon. Tomorrow I'll have news 🙃