@aartigao 2024-04-11
I can confirm that, in our case, the leak was caused due to a misuse of `IO.racePair`.